requirements.txt
numpy
pandas
pre-commit
pytest
faker
pydantic
pytest-mock
pydantic[email]
google-cloud-bigquery
